remove opcar and move docs to car
fix these debugging scripts

fix docs
