+++ selfdrive/car/tests/test_docs.py
@@ -4,7 +4,7 @@ from openpilot.common.basedir import BASEDIR
 from opendbc.car.docs import generate_cars_md, get_all_car_docs
 from openpilot.selfdrive.debug.dump_car_docs import dump_car_docs
 from openpilot.selfdrive.debug.print_docs_diff import print_car_docs_diff
-from openpilot.selfdrive.opcar.docs import CARS_MD_OUT, CARS_MD_TEMPLATE
+from openpilot.selfdrive.car.docs import CARS_MD_OUT, CARS_MD_TEMPLATE
 
 
 class TestCarDocs:
