+++ selfdrive/debug/car/ecu_addrs.py
@@ -2,9 +2,9 @@
 import argparse
 import time
 import cereal.messaging as messaging
+from opendbc.car.ecu_addrs import get_all_ecu_addrs
 from openpilot.common.params import Params
 from openpilot.selfdrive.car.card import can_comm_callbacks, obd_callback
-from openpilot.selfdrive.car.ecu_addrs import get_all_ecu_addrs
 
 
 if __name__ == "__main__":
