+++ selfdrive/debug/car/fw_versions.py
@@ -3,10 +3,10 @@ import time
 import argparse
 import cereal.messaging as messaging
 from cereal import car
+from opendbc.car.fw_versions import get_fw_versions, match_fw_to_car
+from opendbc.car.vin import get_vin
 from openpilot.common.params import Params
 from openpilot.selfdrive.car.card import can_comm_callbacks, obd_callback
-from openpilot.selfdrive.car.fw_versions import get_fw_versions, match_fw_to_car
-from openpilot.selfdrive.car.vin import get_vin
 from typing import Any
 
 Ecu = car.CarParams.Ecu
