--- selfdrive/opcar/__init__.py
@@ -1,1 +1,0 @@
-# This folder will be renamed back to car after the opendbc car split
